feat: add dynamic page title to public profile page (#35)
